Split LGraphGroup
